[feat,meta]: shorten DiscordBotList to DBL and fix feature names
