+++ src/dropwizardWebhooks/java/org/discordbots/webhooks/dropwizard/DBLWebhooks.java
@@ -1,14 +1,15 @@
-package org.discordbots.webhooks.eclipsejetty;
+package org.discordbots.webhooks.dropwizard;
 
 import com.fatboyindustrial.gsonjavatime.OffsetDateTimeConverter;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonIOException;
 import com.google.gson.JsonSyntaxException;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.Context;
+import jakarta.ws.rs.core.Response;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.security.InvalidKeyException;
@@ -26,11 +27,11 @@ import org.discordbots.webhooks.payload.Payload;
 import org.discordbots.webhooks.payload.TestPayload;
 import org.discordbots.webhooks.payload.VoteCreatePayload;
 
-public class DiscordBotListWebhooks extends HttpServlet implements DiscordBotListWebhooksListener {
+public abstract class DBLWebhooks implements DBLWebhooksListener {
   private byte[] secret;
   private final Gson gson;
 
-  public DiscordBotListWebhooks(final String secret) {
+  public DBLWebhooks(final String secret) {
     this.secret = secret.getBytes(StandardCharsets.UTF_8);
     this.gson =
         new GsonBuilder()
@@ -46,10 +47,9 @@ public class DiscordBotListWebhooks extends HttpServlet implements DiscordBotLis
     secret = newSecret.getBytes(StandardCharsets.UTF_8);
   }
 
-  @Override
+  @POST
   @SuppressWarnings("UseSpecificCatch")
-  protected void doPost(HttpServletRequest request, HttpServletResponse response)
-      throws IOException, ServletException {
+  public Response handle(@Context HttpServletRequest request) throws WebApplicationException {
     try {
       final String signatureHeader = request.getHeader("x-topgg-signature");
 
@@ -81,30 +81,28 @@ public class DiscordBotListWebhooks extends HttpServlet implements DiscordBotLis
           hmac.doFinal(String.format("%s.%s", timestamp, body).getBytes(StandardCharsets.UTF_8));
 
       if (!signature.equals(HexFormat.of().formatHex(digest))) {
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        response.getWriter().write("Invalid Authorization");
-
-        return;
+        return Response.status(Response.Status.UNAUTHORIZED)
+            .entity("Invalid Authorization")
+            .build();
       }
 
       final Payload payload = gson.fromJson(body, Payload.class);
       final String trace = request.getHeader("x-topgg-trace");
 
       try {
-        switch (payload.getType()) {
+        return switch (payload.getType()) {
           case "integration.create" ->
-              onIntegrationCreate(
-                  response, payload.getData(gson, IntegrationCreatePayload.class), trace);
+              onIntegrationCreate(payload.getData(gson, IntegrationCreatePayload.class), trace);
           case "integration.delete" ->
-              onIntegrationDelete(
-                  response, payload.getData(gson, IntegrationDeletePayload.class), trace);
-          case "webhook.test" -> onTest(response, payload.getData(gson, TestPayload.class), trace);
-          case "vote.create" ->
-              onVoteCreate(response, payload.getData(gson, VoteCreatePayload.class), trace);
-        }
+              onIntegrationDelete(payload.getData(gson, IntegrationDeletePayload.class), trace);
+          case "webhook.test" -> onTest(payload.getData(gson, TestPayload.class), trace);
+          case "vote.create" -> onVoteCreate(payload.getData(gson, VoteCreatePayload.class), trace);
+          default -> Response.status(Response.Status.BAD_REQUEST).entity("Bad Request").build();
+        };
       } catch (final Throwable ignored) {
-        response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-        response.getWriter().write("Internal Server Error");
+        return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
+            .entity("Internal Server Error")
+            .build();
       }
     } catch (final NoSuchAlgorithmException
         | InvalidKeyException
@@ -114,10 +112,9 @@ public class DiscordBotListWebhooks extends HttpServlet implements DiscordBotLis
         | JsonIOException
         | IOException error) {
       if (error instanceof NoSuchAlgorithmException || error instanceof InvalidKeyException) {
-        throw new ServletException("Unable to find HMAC SHA-256 algorithm", error);
+        throw new WebApplicationException("Unable to find HMAC SHA-256 algorithm", error);
       } else {
-        response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-        response.getWriter().write("Bad Request");
+        return Response.status(Response.Status.BAD_REQUEST).entity("Bad Request").build();
       }
     }
   }
