+++ src/eclipseJettyWebhooks/java/org/discordbots/webhooks/eclipsejetty/DBLWebhooks.java
@@ -1,15 +1,14 @@
-package org.discordbots.webhooks.dropwizard;
+package org.discordbots.webhooks.eclipsejetty;
 
 import com.fatboyindustrial.gsonjavatime.OffsetDateTimeConverter;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonIOException;
 import com.google.gson.JsonSyntaxException;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.ws.rs.POST;
-import jakarta.ws.rs.WebApplicationException;
-import jakarta.ws.rs.core.Context;
-import jakarta.ws.rs.core.Response;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.security.InvalidKeyException;
@@ -27,11 +26,11 @@ import org.discordbots.webhooks.payload.Payload;
 import org.discordbots.webhooks.payload.TestPayload;
 import org.discordbots.webhooks.payload.VoteCreatePayload;
 
-public abstract class DiscordBotListWebhooks implements DiscordBotListWebhooksListener {
+public class DBLWebhooks extends HttpServlet implements DBLWebhooksListener {
   private byte[] secret;
   private final Gson gson;
 
-  public DiscordBotListWebhooks(final String secret) {
+  public DBLWebhooks(final String secret) {
     this.secret = secret.getBytes(StandardCharsets.UTF_8);
     this.gson =
         new GsonBuilder()
@@ -47,9 +46,10 @@ public abstract class DiscordBotListWebhooks implements DiscordBotListWebhooksLi
     secret = newSecret.getBytes(StandardCharsets.UTF_8);
   }
 
-  @POST
+  @Override
   @SuppressWarnings("UseSpecificCatch")
-  public Response handle(@Context HttpServletRequest request) throws WebApplicationException {
+  protected void doPost(HttpServletRequest request, HttpServletResponse response)
+      throws IOException, ServletException {
     try {
       final String signatureHeader = request.getHeader("x-topgg-signature");
 
@@ -81,28 +81,30 @@ public abstract class DiscordBotListWebhooks implements DiscordBotListWebhooksLi
           hmac.doFinal(String.format("%s.%s", timestamp, body).getBytes(StandardCharsets.UTF_8));
 
       if (!signature.equals(HexFormat.of().formatHex(digest))) {
-        return Response.status(Response.Status.UNAUTHORIZED)
-            .entity("Invalid Authorization")
-            .build();
+        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+        response.getWriter().write("Invalid Authorization");
+
+        return;
       }
 
       final Payload payload = gson.fromJson(body, Payload.class);
       final String trace = request.getHeader("x-topgg-trace");
 
       try {
-        return switch (payload.getType()) {
+        switch (payload.getType()) {
           case "integration.create" ->
-              onIntegrationCreate(payload.getData(gson, IntegrationCreatePayload.class), trace);
+              onIntegrationCreate(
+                  response, payload.getData(gson, IntegrationCreatePayload.class), trace);
           case "integration.delete" ->
-              onIntegrationDelete(payload.getData(gson, IntegrationDeletePayload.class), trace);
-          case "webhook.test" -> onTest(payload.getData(gson, TestPayload.class), trace);
-          case "vote.create" -> onVoteCreate(payload.getData(gson, VoteCreatePayload.class), trace);
-          default -> Response.status(Response.Status.BAD_REQUEST).entity("Bad Request").build();
-        };
+              onIntegrationDelete(
+                  response, payload.getData(gson, IntegrationDeletePayload.class), trace);
+          case "webhook.test" -> onTest(response, payload.getData(gson, TestPayload.class), trace);
+          case "vote.create" ->
+              onVoteCreate(response, payload.getData(gson, VoteCreatePayload.class), trace);
+        }
       } catch (final Throwable ignored) {
-        return Response.status(Response.Status.INTERNAL_SERVER_ERROR)
-            .entity("Internal Server Error")
-            .build();
+        response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+        response.getWriter().write("Internal Server Error");
       }
     } catch (final NoSuchAlgorithmException
         | InvalidKeyException
@@ -112,9 +114,10 @@ public abstract class DiscordBotListWebhooks implements DiscordBotListWebhooksLi
         | JsonIOException
         | IOException error) {
       if (error instanceof NoSuchAlgorithmException || error instanceof InvalidKeyException) {
-        throw new WebApplicationException("Unable to find HMAC SHA-256 algorithm", error);
+        throw new ServletException("Unable to find HMAC SHA-256 algorithm", error);
       } else {
-        return Response.status(Response.Status.BAD_REQUEST).entity("Bad Request").build();
+        response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+        response.getWriter().write("Bad Request");
       }
     }
   }
