+++ src/main/java/org/discordbots/api/entity/PaginatedVotes.java
@@ -2,15 +2,14 @@ package org.discordbots.api.entity;
 
 import java.util.List;
 import java.util.concurrent.CompletionStage;
-import org.discordbots.api.DiscordBotListAPI;
+import org.discordbots.api.DBLAPI;
 
 public class PaginatedVotes {
   private final List<Vote> votes;
   private final String cursor;
-  private final DiscordBotListAPI client;
+  private final DBLAPI client;
 
-  public PaginatedVotes(
-      final List<Vote> votes, final String cursor, final DiscordBotListAPI client) {
+  public PaginatedVotes(final List<Vote> votes, final String cursor, final DBLAPI client) {
     this.votes = votes;
     this.cursor = cursor;
     this.client = client;
