+++ src/test/java/org/discordbots/api/DBLWidgetTest.java
@@ -4,28 +4,28 @@ import org.discordbots.api.entity.ProjectType;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 
-public class DiscordBotListWidgetTest {
+public class DBLWidgetTest {
   @ParameterizedTest
   @EnumSource(ProjectType.class)
   public void large(final ProjectType projectType) {
-    DiscordBotListWidget.large(projectType, "123456");
+    DBLWidget.large(projectType, "123456");
   }
 
   @ParameterizedTest
   @EnumSource(ProjectType.class)
   public void votes(final ProjectType projectType) {
-    DiscordBotListWidget.votes(projectType, "123456");
+    DBLWidget.votes(projectType, "123456");
   }
 
   @ParameterizedTest
   @EnumSource(ProjectType.class)
   public void owner(final ProjectType projectType) {
-    DiscordBotListWidget.owner(projectType, "123456");
+    DBLWidget.owner(projectType, "123456");
   }
 
   @ParameterizedTest
   @EnumSource(ProjectType.class)
   public void social(final ProjectType projectType) {
-    DiscordBotListWidget.social(projectType, "123456");
+    DBLWidget.social(projectType, "123456");
   }
 }
